gen-platform/api/agents/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.9.1
python-multipart
pyyaml